WIP: qemu: Use incremented CID if currently in use
The system name is not a guaranteed unique identifier
and may already be in use.

The VHOST_VSOCK_SET_GUEST_CID ioctl may be used
to set the CID before starting qemu
so it's recoverable if it's already in use.

NOTE: This currently doesn't support running as non-root,
which would be easily fixable by binding the cid sooner;
and does not work for ssh, which would require writing the CID
to the output directory and reading it back out when starting ssh.